Move AWSResponse tests into separate file
clean up at same time
