+++ Tests/AWSSDKSwiftCoreTests/AWSClientTests.swift
@@ -23,27 +23,6 @@ import AWSXML
 
 class AWSClientTests: XCTestCase {
 
-    struct AWSHTTPResponseImpl: AWSHTTPResponse {
-        let status: HTTPResponseStatus
-        let headers: HTTPHeaders
-        let body: ByteBuffer?
-
-        init(status: HTTPResponseStatus, headers: HTTPHeaders, body: ByteBuffer?) {
-            self.status = status
-            self.headers = headers
-            self.body = body
-        }
-
-        init(status: HTTPResponseStatus, headers: HTTPHeaders, bodyData: Data?) {
-            var body: ByteBuffer? = nil
-            if let bodyData = bodyData {
-                body = ByteBufferAllocator().buffer(capacity: bodyData.count)
-                body?.writeBytes(bodyData)
-            }
-            self.init(status: status, headers: headers, body: body)
-        }
-    }
-
     struct C: AWSEncodableShape {
         public static var _encoding: [AWSMemberEncoding] = [
              AWSMemberEncoding(label: "value", location: .header(locationName: "value"))
@@ -384,80 +363,6 @@ class AWSClientTests: XCTestCase {
         }
     }
 
-    func testHeaderResponseDecoding() {
-        struct Output: AWSDecodableShape {
-            static let _encoding = [AWSMemberEncoding(label: "h", location: .header(locationName: "header-member"))]
-            let h: String
-            private enum CodingKeys: String, CodingKey {
-                case h = "header-member"
-            }
-        }
-        let response = AWSHTTPResponseImpl(
-            status: .ok,
-            headers: ["header-member": "test-header"],
-            bodyData: nil
-        )
-
-        // XML
-        let queryClient = createAWSClient(serviceProtocol: .query)
-        defer {
-            XCTAssertNoThrow(try queryClient.syncShutdown())
-        }
-        do {
-            let result: Output = try queryClient.validate(operation: "Test", response: response)
-            XCTAssertEqual(result.h, "test-header")
-        } catch {
-            XCTFail("\(error)")
-        }
-
-        // JSON
-        let jsonClient = createAWSClient(serviceProtocol: .restjson)
-        defer {
-            XCTAssertNoThrow(try jsonClient.syncShutdown())
-        }
-        do {
-            let result: Output = try jsonClient.validate(operation: "Test", response: response)
-            XCTAssertEqual(result.h, "test-header")
-        } catch {
-            XCTFail("\(error)")
-        }
-    }
-
-    func testStatusCodeResponseDecoding() {
-        struct Output: AWSDecodableShape {
-            static let _encoding = [AWSMemberEncoding(label: "status", location: .statusCode)]
-            let status: Int
-        }
-        let response = AWSHTTPResponseImpl(
-            status: .ok,
-            headers: HTTPHeaders(),
-            bodyData: nil
-        )
-
-        // XML
-        let xmlClient = createAWSClient(serviceProtocol: .restxml)
-        defer {
-            XCTAssertNoThrow(try xmlClient.syncShutdown())
-        }
-        do {
-            let result: Output = try xmlClient.validate(operation: "Test", response: response)
-            XCTAssertEqual(result.status, 200)
-        } catch {
-            XCTFail("\(error)")
-        }
-        // JSON
-        let jsonClient = createAWSClient(serviceProtocol: .restjson)
-        defer {
-            XCTAssertNoThrow(try jsonClient.syncShutdown())
-        }
-        do {
-            let result: Output = try jsonClient.validate(operation: "Test", response: response)
-            XCTAssertEqual(result.status, 200)
-        } catch {
-            XCTFail("\(error)")
-        }
-    }
-
     func testCreateWithXMLNamespace() {
         struct Input: AWSEncodableShape {
             public static let _xmlNamespace: String? = "https://test.amazonaws.com/doc/2020-03-11/"
@@ -507,245 +412,6 @@ class AWSClientTests: XCTestCase {
         }
     }
 
-    func testValidateXMLResponse() {
-        class Output : AWSDecodableShape {
-            let name : String
-        }
-        let xmlClient = createAWSClient(serviceProtocol: .restxml)
-        defer {
-            XCTAssertNoThrow(try xmlClient.syncShutdown())
-        }
-        let responseBody = "<Output><name>hello</name></Output>"
-        let awsHTTPResponse = AWSHTTPResponseImpl(
-            status: .ok,
-            headers: HTTPHeaders(),
-            bodyData: responseBody.data(using: .utf8)!
-        )
-        do {
-            let output : Output = try xmlClient.validate(operation: "Output", response: awsHTTPResponse)
-            XCTAssertEqual(output.name, "hello")
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
-    func testValidateXMLCodablePayloadResponse() {
-        class Output : AWSDecodableShape & AWSShapeWithPayload {
-            static let _encoding = [AWSMemberEncoding(label: "contentType", location: .header(locationName: "content-type"))]
-            static let _payloadPath: String = "name"
-            let name : String
-            let contentType: String
-
-            private enum CodingKeys: String, CodingKey {
-                case name = "name"
-                case contentType = "content-type"
-            }
-        }
-        let xmlClient = createAWSClient(serviceProtocol: .restxml)
-        defer {
-            XCTAssertNoThrow(try xmlClient.syncShutdown())
-        }
-let response = AWSHTTPResponseImpl(
-            status: .ok,
-            headers: ["Content-Type":"application/json"],
-            bodyData: "<name>hello</name>".data(using: .utf8)!
-        )
-        do {
-            let output : Output = try xmlClient.validate(operation: "Output", response: response)
-            XCTAssertEqual(output.name, "hello")
-            XCTAssertEqual(output.contentType, "application/json")
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
-    func testValidateXMLRawPayloadResponse() {
-        class Output : AWSDecodableShape, AWSShapeWithPayload {
-            static let _payloadPath: String = "body"
-            static let _payloadOptions: PayloadOptions = .raw
-            let body : AWSPayload
-        }
-        let xmlClient = createAWSClient(serviceProtocol: .restxml)
-        defer {
-            XCTAssertNoThrow(try xmlClient.syncShutdown())
-        }
-        let response = AWSHTTPResponseImpl(
-            status: .ok,
-            headers: HTTPHeaders(),
-            bodyData: "{\"name\":\"hello\"}".data(using: .utf8)!
-        )
-        do {
-            let output : Output = try xmlClient.validate(operation: "Output", response: response)
-            XCTAssertEqual(output.body.asData(), "{\"name\":\"hello\"}".data(using: .utf8))
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
-    func testXMLError() {
-        let response = AWSHTTPResponseImpl(
-            status: .notFound,
-            headers: HTTPHeaders(),
-            bodyData: "<Error><Code>NoSuchKey</Code><Message>It doesn't exist</Message></Error>".data(using: .utf8)!
-        )
-        let xmlClient = createAWSClient(serviceProtocol: .restxml, possibleErrorTypes: [ServiceErrorType.self])
-        defer {
-            XCTAssertNoThrow(try xmlClient.syncShutdown())
-        }
-        let error = xmlClient.createError(for: response)
-        if case ServiceErrorType.noSuchKey(let message) = error {
-            XCTAssertEqual(message, "It doesn't exist")
-        } else {
-            XCTFail("Error is not noSuchKey")
-        }
-    }
-
-    func testQueryError() {
-        let response = AWSHTTPResponseImpl(
-            status: .notFound,
-            headers: HTTPHeaders(),
-            bodyData: "<ErrorResponse><Error><Code>MessageRejected</Code><Message>Don't like it</Message></Error></ErrorResponse>".data(using: .utf8)!
-        )
-        let queryClient = createAWSClient(serviceProtocol: .query, possibleErrorTypes: [ServiceErrorType.self])
-        defer {
-            XCTAssertNoThrow(try queryClient.syncShutdown())
-        }
-        let error = queryClient.createError(for: response)
-        if case ServiceErrorType.messageRejected(let message) = error {
-            XCTAssertEqual(message, "Don't like it")
-        } else {
-            XCTFail("Creating the wrong error")
-        }
-    }
-
-
-    func testValidateJSONResponse() {
-        class Output : AWSDecodableShape {
-            let name : String
-        }
-        let response = AWSHTTPResponseImpl(
-            status: .ok,
-            headers: HTTPHeaders(),
-            bodyData: "{\"name\":\"hello\"}".data(using: .utf8)!
-        )
-        let jsonClient = createAWSClient(serviceProtocol: .json(version: "1.1"))
-        defer {
-            XCTAssertNoThrow(try jsonClient.syncShutdown())
-        }
-        do {
-            let output : Output = try jsonClient.validate(operation: "Output", response: response)
-            XCTAssertEqual(output.name, "hello")
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
-    func testValidateJSONCodablePayloadResponse() {
-        class Output2 : AWSDecodableShape {
-            let name : String
-        }
-        struct Output : AWSDecodableShape & AWSShapeWithPayload {
-            static let _payloadPath: String = "output2"
-            let output2 : Output2
-        }
-        let jsonClient = createAWSClient(serviceProtocol: .json(version: "1.1"))
-        defer {
-            XCTAssertNoThrow(try jsonClient.syncShutdown())
-        }
-        let response = AWSHTTPResponseImpl(
-            status: .ok,
-            headers: HTTPHeaders(),
-            bodyData: "{\"name\":\"hello\"}".data(using: .utf8)!
-        )
-        do {
-            let output : Output = try jsonClient.validate(operation: "Output", response: response)
-            XCTAssertEqual(output.output2.name, "hello")
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
-    func testValidateJSONRawPayloadResponse() {
-        struct Output : AWSDecodableShape, AWSShapeWithPayload {
-            static let _payloadPath: String = "body"
-            static let _payloadOptions: PayloadOptions = .raw
-            public static var _encoding = [
-                AWSMemberEncoding(label: "contentType", location: .header(locationName: "content-type")),
-            ]
-            let body : AWSPayload
-        }
-        let jsonClient = createAWSClient(serviceProtocol: .json(version: "1.1"))
-        defer {
-            XCTAssertNoThrow(try jsonClient.syncShutdown())
-        }
-        let response = AWSHTTPResponseImpl(
-            status: .ok,
-            headers: ["Content-Type":"application/json"],
-            bodyData: "{\"name\":\"hello\"}".data(using: .utf8)!
-        )
-        do {
-            let output : Output = try jsonClient.validate(operation: "Output", response: response)
-            XCTAssertEqual(output.body.asData(), "{\"name\":\"hello\"}".data(using: .utf8))
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
-    func testJSONError() {
-        let response = AWSHTTPResponseImpl(
-            status: .notFound,
-            headers: HTTPHeaders(),
-            bodyData: "{\"__type\":\"ResourceNotFoundException\", \"message\": \"Donald Where's Your Troosers?\"}".data(using: .utf8)!
-        )
-        let jsonClient = createAWSClient(serviceProtocol: .json(version: "1.1"), possibleErrorTypes: [ServiceErrorType.self])
-        defer {
-            XCTAssertNoThrow(try jsonClient.syncShutdown())
-        }
-        let error = jsonClient.createError(for: response)
-        if case ServiceErrorType.resourceNotFoundException(let message) = error {
-            XCTAssertEqual(message, "Donald Where's Your Troosers?")
-        } else {
-            XCTFail("Error is not resourceNotFoundException")
-        }
-    }
-
-
-    func testProcessHAL() {
-        struct Output : AWSDecodableShape {
-            let s: String
-            let i: Int
-        }
-        struct Output2 : AWSDecodableShape {
-            let a: [Output]
-            let d: Double
-            let b: Bool
-        }
-        let jsonClient = createAWSClient(serviceProtocol: .json(version: "1.1"))
-        defer {
-            XCTAssertNoThrow(try jsonClient.syncShutdown())
-        }
-        var buffer = ByteBufferAllocator().buffer(capacity: 0)
-        buffer.writeString(
-            """
-            {"_embedded": {"a": [{"s":"Hello", "i":1234}, {"s":"Hello2", "i":12345}]}, "d":3.14, "b":true}
-            """
-        )
-        let response = HTTPClient.Response(
-            host: "localhost",
-            status: .ok,
-            headers: ["Content-Type":"application/hal+json"],
-            body: buffer
-        )
-        do {
-            let output : Output2 = try jsonClient.validate(operation: "Output", response: response)
-            XCTAssertEqual(output.a.count, 2)
-            XCTAssertEqual(output.d, 3.14)
-            XCTAssertEqual(output.a[1].s, "Hello2")
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
     func testDataInJsonPayload() {
         struct DataContainer: AWSEncodableShape {
             let data: Data
@@ -766,32 +432,6 @@ let response = AWSHTTPResponseImpl(
                 httpMethod: "POST",
                 input: input
             )
-        } catch {
-            XCTFail(error.localizedDescription)
-        }
-    }
-
-    func testPayloadDataInResponse() {
-        struct Response: AWSDecodableShape, AWSShapeWithPayload {
-            public static let _payloadPath: String = "payload"
-            public static let _payloadOptions: PayloadOptions = .raw
-            let payload: AWSPayload
-        }
-        let jsonClient = createAWSClient(serviceProtocol: .json(version: "1.1"))
-        defer {
-            XCTAssertNoThrow(try jsonClient.syncShutdown())
-        }
-        var buffer = ByteBufferAllocator().buffer(capacity: 0)
-        buffer.writeString("TestString")
-        let response = HTTPClient.Response(
-            host: "localhost",
-            status: .ok,
-            headers: ["Content-Type":"application/hal+json"],
-            body: buffer
-        )
-        do {
-            let output : Response = try jsonClient.validate(operation: "Output", response: response)
-            XCTAssertEqual(output.payload.asString(), "TestString")
         } catch {
             XCTFail(error.localizedDescription)
         }
@@ -924,24 +564,6 @@ let response = AWSHTTPResponseImpl(
             XCTAssertEqual(request.body.asString(), "Action=Test&Array.1=entry1&Array.2=entry2&Version=2013-12-02")
         } catch {
             XCTFail("\(error)")
-        }
-    }
-
-    func testEC2ValidateError() {
-        let response = AWSHTTPResponseImpl(
-            status: .notFound,
-            headers: HTTPHeaders(),
-            bodyData: "<Errors><Error><Code>NoSuchKey</Code><Message>It doesn't exist</Message></Error></Errors>".data(using: .utf8)!
-        )
-        let client = createAWSClient(serviceProtocol: .ec2)
-        defer {
-            XCTAssertNoThrow(try client.syncShutdown())
-        }
-        if let error = client.createError(for: response) as? AWSResponseError {
-            XCTAssertEqual(error.errorCode, "NoSuchKey")
-            XCTAssertEqual(error.message, "It doesn't exist")
-        } else {
-            XCTFail("Throwing the wrong error")
         }
     }
 
@@ -1399,40 +1021,5 @@ let response = AWSHTTPResponseImpl(
             XCTFail("Unexpected error: \(error)")
         }
 
-    }
-}
-
-
-
-/// Error enum for Kinesis
-public enum ServiceErrorType: AWSErrorType {
-    case resourceNotFoundException(message: String?)
-    case noSuchKey(message: String?)
-    case messageRejected(message: String?)
-}
-
-extension ServiceErrorType {
-    public init?(errorCode: String, message: String?){
-        switch errorCode {
-        case "ResourceNotFoundException":
-            self = .resourceNotFoundException(message: message)
-        case "NoSuchKey":
-            self = .noSuchKey(message: message)
-        case "MessageRejected":
-            self = .messageRejected(message: message)
-        default:
-            return nil
-        }
-    }
-
-    public var description : String {
-        switch self {
-        case .resourceNotFoundException(let message):
-            return "ResourceNotFoundException :\(message ?? "")"
-        case .noSuchKey(let message):
-            return "NoSuchKey :\(message ?? "")"
-        case .messageRejected(let message):
-            return "MessageRejected :\(message ?? "")"
-        }
     }
 }
